requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
black==19.3b0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
coverage==4.5.4
cryptography==2.7
cssselect==1.1.0
decorator==4.4.0
Deprecated==1.2.6
entrypoints==0.3
future==0.17.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.20
incremental==17.5.0
ipython==7.6.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.14.1
jeepney==0.4.1
Jinja2==2.10.1
keyring==19.1.0
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
MouseInfo==0.0.4
nose2==0.9.1
numpy==1.17.1
oauthlib==3.1.0
packaging==19.1
parsel==1.5.2
parso==0.5.1
pendulum==2.0.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkg-resources==0.0.0
pluggy==0.12.0
pprint==0.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
py-cpuinfo==5.0.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
PyAutoGUI==0.9.47
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.1.1
PyGetWindow==0.0.7
pygit2==0.28.2
PyGithub==1.43.8
pyglet==1.4.2
pyglet-ffmpeg2==0.1.12
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMsgBox==1.0.7
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
PySocks==1.7.0
pytest==5.1.2
pytest-benchmark==3.2.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python3-xlib==0.15
pytiled-parser==0.9.0
PyTweening==1.0.3
pytz==2019.2
pytzdata==2019.2
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
Scrapy==1.7.3
SecretStorage==3.1.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
termcolor==1.1.0
toml==0.10.0
traitlets==4.3.2
tweepy==3.8.0
Twisted==19.7.0
ujson==1.35
urllib3==1.25.3
uvloop==0.13.0
vibora==0.0.6
w3lib==1.21.0
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
yagmail==0.11.220
zipp==0.6.0
zope.interface==4.6.0
